refactor: use image value as display type instead of tag for smart app compatibility
N
