+++ apiserver/paasng/paasng/platform/engine/constants.py
@@ -174,13 +174,3 @@ class NoPrefixAppRunTimeBuiltinEnv(str, StructuredEnum):
     """Built-in envs without prefix in the app runtime"""
 
     PORT = EnumField("PORT", label=_("目标端口号，值为 5000"))
-
-
-class VersionType(str, StructuredEnum):
-    """版本类型. 对应 VersionInfo.version_type"""
-
-    TAG = EnumField("tag", label="用于 Git 仓库、云原生镜像应用、旧镜像应用、镜像模式的 S-Mart 应用")
-    BRANCH = EnumField("branch", label="用于 SVN 仓库、Git 仓库")
-    TRUNK = EnumField("trunk", label="用于 SVN 仓库")
-    IMAGE = EnumField("image", label="仅用于云原生应用选择已构建的镜像部署时")
-    PACKAGE = EnumField("package", label="用于 lesscode 应用和二进制的 S-Mart 应用")
