+++ apiserver/paasng/paasng/platform/engine/deploy/start.py
@@ -22,7 +22,7 @@ from typing import Dict, Optional
 from paas_wl.bk_app.cnative.specs.models import AppModelResource
 from paasng.platform.applications.constants import ApplicationType
 from paasng.platform.applications.models import ModuleEnvironment
-from paasng.platform.engine.constants import OperationTypes, RuntimeType, VersionType
+from paasng.platform.engine.constants import OperationTypes, RuntimeType
 from paasng.platform.engine.deploy.building import start_build, start_build_error_callback
 from paasng.platform.engine.deploy.image_release import release_without_build
 from paasng.platform.engine.models.deployment import Deployment
@@ -33,6 +33,7 @@ from paasng.platform.modules.constants import SourceOrigin
 from paasng.platform.modules.models import Module
 from paasng.platform.modules.models.deploy_config import Hook, HookList
 from paasng.platform.modules.specs import ModuleSpecs
+from paasng.platform.sourcectl.constants import VersionType
 from paasng.platform.sourcectl.models import VersionInfo
 from paasng.platform.sourcectl.version_services import get_version_service
