+++ apiserver/paasng/paasng/platform/engine/management/commands/deploy_bkapp.py
@@ -33,24 +33,25 @@ from django.core.management.base import BaseCommand, CommandError
 
 from paasng.core.core.storages.redisdb import get_default_redis
 from paasng.platform.applications.models import Application
-from paasng.platform.engine.constants import JobStatus, VersionType
+from paasng.platform.engine.constants import JobStatus
 from paasng.platform.engine.deploy.start import DeployTaskRunner, initialize_deployment
 from paasng.platform.engine.models.deployment import Deployment
 from paasng.platform.engine.utils.output import Style
 from paasng.platform.engine.workflow import DeploymentCoordinator, ServerSendEvent
+from paasng.platform.sourcectl.constants import VersionType
 from paasng.platform.sourcectl.models import VersionInfo
 from paasng.platform.sourcectl.version_services import get_version_service
 from paasng.utils.error_codes import error_codes
 
 REVISION_HELP_TEXT = """\
 revision 描述需要部署的版本.
-对于 git/svn 应用, 支持部署 branch or tag, 对于 s-mart 应用, 支持 package or tag.
+对于 git/svn 应用, 支持部署 branch or tag, 对于 s-mart 应用, 支持 package or image.
 
 Example:
 git/svn tag: 'tag:你的标签'
 git/svn branch: 'branch:你的分支名'
 s-mart(二进制): 'package:源码包版本号'
-s-mart(镜像): 'tag:镜像 tag'
+s-mart(镜像): 'image:镜像 tag'
 """
 logger = logging.getLogger("commands")
 
@@ -110,8 +111,10 @@ class Command(BaseCommand):
         version_service = get_version_service(module, operator=operator.pk)
 
         version_type, version_name = smart_revision.split(":", 1)
-        if version_type == VersionType.IMAGE.value:
-            raise DeployError("不支持 --revision 以 image: 开头的版本")
+
+        # 针对 s-mart 镜像应用, 后端部署接口调整为仅支持 version_type 值为 tag, 因此这里将 image 转换为 tag
+        if version_type == VersionType.IMAGE.value and application.is_smart_app:
+            version_type = VersionType.TAG.value
 
         revision = version_service.extract_smart_revision(smart_revision)
         version_info = VersionInfo(revision, version_name, version_type)
