+++ apiserver/paasng/paasng/platform/sourcectl/serializers.py
@@ -21,6 +21,7 @@ import logging
 from rest_framework import serializers
 from rest_framework.validators import UniqueTogetherValidator
 
+from paasng.platform.sourcectl.constants import VersionType
 from paasng.platform.sourcectl.models import RepositoryInstance, SvnAccount, SvnRepository
 from paasng.platform.sourcectl.source_types import get_sourcectl_type
 from paasng.platform.sourcectl.type_specs import BkSvnSourceTypeSpec
@@ -117,12 +118,21 @@ class SvnAccountCreateSLZ(serializers.ModelSerializer):
 class AlternativeVersionSLZ(serializers.Serializer):
     name = serializers.CharField()
     type = serializers.CharField()
+    display_type = serializers.SerializerMethodField()
     revision = serializers.CharField()
     url = serializers.CharField()
     last_update = serializers.DateTimeField()
     message = serializers.CharField(help_text="tag description or commit message")
 
     extra = serializers.JSONField(default=dict)
+
+    def get_display_type(self, obj):
+        # smart 保持前端展示为 image. 相关背景 pr:
+        # https://github.com/TencentBlueKing/blueking-paas/pull/1306/files
+        # https://github.com/TencentBlueKing/blueking-paas/pull/1308/files
+        if self.context.get("is_smart_app") and obj.type == VersionType.TAG.value:
+            return VersionType.IMAGE.value
+        return obj.type
 
 
 class PageNumberPaginationSLZ(serializers.Serializer):
