+++ apiserver/paasng/tests/api/test_engine.py
@@ -26,11 +26,11 @@ from django.urls import reverse
 from django_dynamic_fixture import G
 
 from paasng.platform.applications.constants import ApplicationRole
-from paasng.platform.engine.constants import VersionType
 from paasng.platform.engine.models.deployment import Deployment
 from paasng.platform.engine.workflow import DeploymentCoordinator
 from paasng.platform.environments.constants import EnvRoleOperation
 from paasng.platform.environments.models import EnvRoleProtection
+from paasng.platform.sourcectl.constants import VersionType
 
 logger = logging.getLogger(__name__)
 pytestmark = pytest.mark.django_db
